perf(cms): share cached tenant-by-api-key lookup in access control
